COM-2989: use featherType and move header
