+++ src/components/CourseProfileHeader/styles.ts
@@ -1,7 +1,7 @@
 import { StyleSheet } from 'react-native';
-import { WHITE, GREY } from '../../../styles/colors';
-import { MAIN_MARGIN_LEFT } from '../../../styles/metrics';
-import { FIRA_SANS_BLACK } from '../../../styles/fonts';
+import { WHITE, GREY } from '../../styles/colors';
+import { MAIN_MARGIN_LEFT } from '../../styles/metrics';
+import { FIRA_SANS_BLACK } from '../../styles/fonts';
 
 const imageHeight = 200;
 const styles = StyleSheet.create({
